+++ database/migrations/2021_01_07_152616_create_reports_table.php
@@ -23,8 +23,8 @@ class CreateReportsTable extends Migration
             $table->enum('status', ['ACTIVO', 'INACTIVO']); 
             $table->timestamps();
             $table->softDeletes($column = 'deleted_at', $precision = 0);
-            $table->unsignedInteger('fk_notice_types');
-            $table->unsignedInteger('fk_users');
+            $table->unsignedBigInteger('fk_notice_types');
+            $table->unsignedBigInteger('fk_users');
 
             $table->foreign('fk_notice_types')->references('id')->on('notice_types');
             $table->foreign('fk_users')->references('id')->on('users');
